requirements.txt
python-barcode==0.14.0
Pillow==10.0.0
pywin32==306